coding the analog in task in a more flexible way
Now analog in channels can be declared and configured using the
ServerSettings object editor.

The analog in task can also be triggered by an external source, allowing
for good synchronization even when a variable timebase is used.

PS : for some reason, to have atticus compiled on my computer, I had to
move the "Opal Kelly" directory into the "AtticusServer" one. I do not
commit this change, but this is weird.